requirements.txt
alembic
bcrypt
blinker
certifi
cffi
chardet
click
Flask
Flask-Bcrypt
Flask-Dance
Flask-DebugToolbar
Flask-Login
Flask-Mail
Flask-Migrate
Flask-OAuth
Flask-OAuthlib
Flask-SQLAlchemy
Flask-WTF
httplib2
idna
itsdangerous
Jinja2
lazy
Mako
MarkupSafe
oauth2
oauthlib
Pillow
pycparser
python-dateutil
python-editor
requests
requests-oauthlib
six
SQLAlchemy
SQLAlchemy-Utils
urllib3
URLObject
Werkzeug
wincertstore
WTForms